requirements.txt
openai>=1.0.0
pypdf2>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
tiktoken>=0.5.0
rich>=13.0.0
python-dotenv>=1.0.0
